feat: playback settings
