+++ src/components/setting/developer/DeveloperSettings.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import { View, ScrollView, Alert } from 'react-native';
 import { List } from 'react-native-paper';
 import { useTranslation } from 'react-i18next';
-import { useStore } from 'zustand';
 
 // eslint-disable-next-line import/no-unresolved
 import { APPSTORE, LOCKHASH } from '@env';
@@ -14,8 +13,6 @@ import useVersionCheck from '@hooks/useVersionCheck';
 import { SelectSettingEntry, SettingEntry } from '../helpers/SettingEntry';
 import NoxCache from '@utils/Cache';
 import useCleanCache from '@hooks/useCleanCache';
-import appStore from '@stores/appStore';
-import { saveFadeInterval } from '@utils/ChromeStorage';
 import GroupView from '../../background/GroupView';
 import showLog from '../debug/Log';
 import { showDebugLog } from '../debug/DebugConsole';
@@ -29,31 +26,12 @@ import SelectSetting from '../helpers/SelectSetting';
 import { SelectDialogChildren } from '../SelectDialogWrapper';
 import { Route, Icons } from './enums';
 import OTA from './OTA';
-import EqualizerButton from '../equalizer/EQButton';
 
-const FadeOptions = [0, 250, 500, 1000];
-const CrossFadeOptions = [0, 2500, 5000, 7500, 12000];
 const ArtworkResOptions = [0, 240, 360, 480, 720, 1080];
 
 const developerSettings: { [key: string]: SettingEntry } = {
-  noInterruption: {
-    settingName: 'noInterruption',
-    settingCategory: 'DeveloperSettings',
-  },
-  prefetchTrack: {
-    settingName: 'prefetchTrack',
-    settingCategory: 'GeneralSettings',
-  },
-  audioOffload: {
-    settingName: 'audioOffload',
-    settingCategory: 'GeneralSettings',
-  },
   parseEmbeddedArtwork: {
     settingName: 'parseEmbeddedArtwork',
-    settingCategory: 'GeneralSettings',
-  },
-  skipSilence: {
-    settingName: 'skipSilence',
     settingCategory: 'GeneralSettings',
   },
   memoryEfficiency: {
@@ -89,7 +67,6 @@ export const Home = ({
   const { t } = useTranslation();
   const { checkVersion } = useVersionCheck();
   const { orphanedCache, cleanOrphanedCache } = useCleanCache();
-  const fadeIntervalMs = useStore(appStore, state => state.fadeIntervalMs);
 
   // HACK: doesnt render on my phone?!
   const logLevelString = [
@@ -120,38 +97,6 @@ export const Home = ({
         setSelectVisible(false);
       },
       title: t('DeveloperSettings.LogLevelName'),
-    } as SelectSettingEntry<number>);
-  };
-
-  const selectFade = () => {
-    setSelectVisible(true);
-    setCurrentSelectOption({
-      options: FadeOptions,
-      renderOption: (option: number) => String(option),
-      defaultIndex: 0,
-      onClose: () => setSelectVisible(false),
-      onSubmit: (index: number) => {
-        saveFadeInterval(FadeOptions[index]).then(selfDestruct);
-        setSelectVisible(false);
-      },
-      title: t('DeveloperSettings.FadeTitle'),
-    } as SelectSettingEntry<number>);
-  };
-
-  const selectCrossFade = () => {
-    setSelectVisible(true);
-    setCurrentSelectOption({
-      options: CrossFadeOptions,
-      renderOption: (option: number) => String(option),
-      defaultIndex: 0,
-      onClose: () => setSelectVisible(false),
-      onSubmit: (index: number) => {
-        setPlayerSetting({ crossfade: CrossFadeOptions[index] / 1000 }).then(
-          selfDestruct,
-        );
-        setSelectVisible(false);
-      },
-      title: t('DeveloperSettings.crossfadeTitle'),
     } as SelectSettingEntry<number>);
   };
 
@@ -230,23 +175,6 @@ export const Home = ({
             `${val}: ${logLevelString[getState().logLevel]}`
           }
         />
-        {isAndroid && <EqualizerButton />}
-        <SettingListItem
-          icon={Icons.fade}
-          settingName="Fade"
-          onPress={selectFade}
-          settingCategory="DeveloperSettings"
-          modifyDescription={val => `${val}: ${fadeIntervalMs}ms`}
-        />
-        {isAndroid && (
-          <SettingListItem
-            icon={Icons.crossfade}
-            settingName="crossfade"
-            onPress={selectCrossFade}
-            settingCategory="DeveloperSettings"
-            modifyDescription={v => `${v}: ${playerSetting.crossfade * 1000}ms`}
-          />
-        )}
         <SelectSetting
           setVisible={setSelectVisible}
           setCurrentOption={setCurrentSelectOption}
